chore: tests series only refactor
