+++ tests/expr_and_series/pipe_test.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-import narwhals as nw
+import narwhals.stable.v1 as nw
 from tests.utils import compare_dicts
 
 input_list = {"a": [2, 4, 6, 8]}
