+++ tests/expr_and_series/unary_test.py
@@ -2,7 +2,7 @@ from typing import Any
 
 import pytest
 
-import narwhals as nw
+import narwhals.stable.v1 as nw
 from tests.utils import compare_dicts
