--- tests/series_only/arithmetic_test.py
@@ -1,98 +1,0 @@
-from __future__ import annotations
-
-from typing import Any
-
-import hypothesis.strategies as st
-import pandas as pd
-import polars as pl
-import pyarrow as pa
-import pytest
-from hypothesis import assume
-from hypothesis import given
-
-import narwhals.stable.v1 as nw
-from narwhals.utils import parse_version
-from tests.utils import compare_dicts
-
-
-def test_truediv_same_dims(constructor_eager: Any, request: Any) -> None:
-    if "polars" in str(constructor_eager):
-        # https://github.com/pola-rs/polars/issues/17760
-        request.applymarker(pytest.mark.xfail)
-    s_left = nw.from_native(constructor_eager({"a": [1, 2, 3]}), eager_only=True)["a"]
-    s_right = nw.from_native(constructor_eager({"a": [2, 2, 1]}), eager_only=True)["a"]
-    result = (s_left / s_right).to_list()
-    assert result == [0.5, 1.0, 3.0]
-    result = (s_left.__rtruediv__(s_right)).to_list()
-    assert result == [2, 1, 1 / 3]
-
-
-@pytest.mark.slow()
-@given(  # type: ignore[misc]
-    left=st.integers(-100, 100),
-    right=st.integers(-100, 100),
-)
-@pytest.mark.skipif(
-    parse_version(pd.__version__) < (2, 0), reason="convert_dtypes not available"
-)
-def test_floordiv(left: int, right: int) -> None:
-    # hypothesis complains if we add `constructor` as an argument, so this
-    # test is a bit manual unfortunately
-    assume(right != 0)
-    expected = {"a": [left // right]}
-    result = nw.from_native(pd.DataFrame({"a": [left]}), eager_only=True).select(
-        nw.col("a") // right
-    )
-    compare_dicts(result, expected)
-    if parse_version(pd.__version__) < (2, 2):  # pragma: no cover
-        # Bug in old version of pandas
-        pass
-    else:
-        result = nw.from_native(
-            pd.DataFrame({"a": [left]}).convert_dtypes(dtype_backend="pyarrow"),
-            eager_only=True,
-        ).select(nw.col("a") // right)
-        compare_dicts(result, expected)
-    result = nw.from_native(
-        pd.DataFrame({"a": [left]}).convert_dtypes(), eager_only=True
-    ).select(nw.col("a") // right)
-    compare_dicts(result, expected)
-    result = nw.from_native(pl.DataFrame({"a": [left]}), eager_only=True).select(
-        nw.col("a") // right
-    )
-    compare_dicts(result, expected)
-    result = nw.from_native(pa.table({"a": [left]}), eager_only=True).select(
-        nw.col("a") // right
-    )
-    compare_dicts(result, expected)
-
-
-@pytest.mark.slow()
-@given(  # type: ignore[misc]
-    left=st.integers(-100, 100),
-    right=st.integers(-100, 100),
-)
-@pytest.mark.skipif(
-    parse_version(pd.__version__) < (2, 0), reason="convert_dtypes not available"
-)
-def test_mod(left: int, right: int) -> None:
-    # hypothesis complains if we add `constructor` as an argument, so this
-    # test is a bit manual unfortunately
-    assume(right != 0)
-    expected = {"a": [left % right]}
-    result = nw.from_native(pd.DataFrame({"a": [left]}), eager_only=True).select(
-        nw.col("a") % right
-    )
-    compare_dicts(result, expected)
-    result = nw.from_native(
-        pd.DataFrame({"a": [left]}).convert_dtypes(), eager_only=True
-    ).select(nw.col("a") % right)
-    compare_dicts(result, expected)
-    result = nw.from_native(pl.DataFrame({"a": [left]}), eager_only=True).select(
-        nw.col("a") % right
-    )
-    compare_dicts(result, expected)
-    result = nw.from_native(pa.table({"a": [left]}), eager_only=True).select(
-        nw.col("a") % right
-    )
-    compare_dicts(result, expected)
